# swagger/portal-api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: wicked.haufe.io Portal API
  description: The API which the API Portal uses to store and retrieve information
 
# Under securityDefinitions we declare which security schemes can be used.
# These definitions do not affect to the operations, but can be applied
# using the 'security' key at the global level or the operation level.
# In this sample, we'll see how to apply the security definition globally,
# and override it at the operation level.
#
# Note that for Access Code Flow, you must provide the Token URL.
securityDefinitions:
  # For personalized access to data
  accessCode:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/auth/local/portal-api/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/auth/local/portal-api/oauth2/authorize
    flow: accessCode
    scopes:
      read_plans: Read plans
      read_groups: Read user group definitions
      read_content: Read custom/markdown/HTML content
      health: Read out health information
      read_templates: Read templates
      read_auth_servers: Read authorization servers
      read_apis: Read API definitions
      read_users: Read user information
      write_users: Write user information
      read_applications: Read application data
      write_applications: Write application data
      read_subscriptions: Read subscriptions
      write_subscriptions: Write/edit/delete subscriptions
      read_approvals: Read pending approvals
      read_verifications: Read email and password verifications
      write_verifications: Create or delete verifications
      read_registrations: Read user registrations
      write_registrations: Write user registrations
      read_grants: Read user API grants
      write_grants: Write user API grants
      webhooks: Access and change webhooks
  
  implicit:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/portal-api/v1/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/portal-auth/api/portal-api
    flow: accessCode
    scopes:
      read_plans: Read plans
...
produces:
  - application/hal+json
host: 'localhost:3001'
basePath: /portal-api/v1
...
paths:
...
  '/users/{userId}':
    get:
      security:
      - accessCode: ['read_users']
      - implicit: ['read_users']
      - password: ['read_users']
      tags:
      - users
      parameters:
        - name: userId
          in: path
          description: The ID of the user to retrieve
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/userType'
        '403':
          description: Not allowed
    patch:
      security:
      - accessCode: ['write_users']
      - implicit: ['write_users']
      - password: ['write_users']
      - clientCredentials: []
      tags:
        - users
      parameters:
        - name: userId
          in: path
          description: The ID of the user to patch
          type: string
          required: true
        - name: user
          description: The user data to patch
          in: body
          schema:
            $ref: '#/definitions/userCreateType'
          required: true
        - name: X-VerificationId
          in: header
          description: 'A verification ID, this is used for updating lost passwords'
          type: string
          required: false
      responses:
        '200':
          description: Patched
          schema:
            $ref: '#/definitions/userType'
        '403':
          description: Not allowed
    delete:
      security:
      - accessCode: ['write_users']
      - implicit: ['write_users']
      - password: ['write_users']
      tags:
        - users
      parameters:
        - name: userId
          in: path
          description: The ID of the user to delete
          type: string
          required: true
      responses:
        '204':
          description: Deleted
        '403':
          description: Not allowed
...
  '/applications':
    get:
      description: Retrieve a list of all applications (admin)
      security:
      - accessCode: ['read_applications']
      - implicit: ['read_applications']
      - password: ['read_applications']
      tags:
        - applications
      responses:
        '200':
          description: Successful response
          schema:
            title: ApplicationIdArray
            type: array
            items:
              title: AppId
              type: object
              properties:
                id:
                  type: string
        '403':
          description: Not allowed
    post:
      security:
      - accessCode: ['write_applications']
      - implicit: ['write_applications']
      - password: ['write_applications']
      tags:
        - applications
      parameters:
        - name: application
          in: body
          schema:
            $ref: '#/definitions/applicationCreateType'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/applicationType'
  '/application/roles':
    get:
      security:
      - accessCode: ['read_applications']
      - implicit: ['read_applications']
      - password: ['read_applications']
      tags:
        - applications
      description: Returns a static list of user roles for applications
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              title: RoleDescription
              type: object
              properties:
                role:
                  type: string
                desc:
                  type: string
  '/applications/{appId}':
    get:
      security:
      - accessCode: ['read_applications']
      - implicit: ['read_applications']
      - password: ['read_applications']
      tags:
        - applications
      parameters:
        - name: appId
          in: path
          description: ID of Application to retrieve
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/applicationType'
        '403':
          description: Not allowed
    patch:
      security:
      - accessCode: ['write_applications']
      - implicit: ['write_applications']
      - password: ['write_applications']
      tags:
        - applications
      parameters:
        - name: appId
          in: path
          description: ID of Application to create or update
          type: string
          required: true
        - name: application
          in: body
          description: Application entity to patch
          schema:
            $ref: '#/definitions/applicationCreateType'
          required: true
      responses:
        '200':
          description: 'Success, patched'
          schema:
            $ref: '#/definitions/applicationType'
        '400':
          description: Bad request
        '403':
          description: Not allowed
    delete:
      security:
      - accessCode: ['write_applications']
      - implicit: ['write_applications']
      - password: ['write_applications']
      tags:
        - applications
      parameters:
        - name: appId
          in: path
          description: ID of Application to delete
          type: string
          required: true
      responses:
        '204':
          description: 'Deleted, no content'
        '403':
          description: Not allowed
  '/applications/{appId}/owners':
    post:
      security:
      - accessCode: ['write_applications']
      - implicit: ['write_applications']
      - password: ['write_applications']
      tags:
        - applications
      description: Adds an Owner to an application
      parameters:
        - name: appId
          in: path
          description: ID of Application to delete
          type: string
          required: true
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              role:
                type: string
                enum:
                  - admin
                  - collaborator
                  - reader
      responses:
        '200':
          description: 'OK, owner was already present'
          schema:
            $ref: '#/definitions/applicationType'
        '201':
          description: Successfully added owner
          schema:
            $ref: '#/definitions/applicationType'
        '400':
          description: 'Bad request, i.e. email not found'
        '403':
          description: Not allowed
    delete:
      security:
      - accessCode: ['write_applications']
      - implicit: ['write_applications']
      - password: ['write_applications']
      tags:
        - applications
      description: Removes an owner from an application
      parameters:
        - name: appId
          in: path
          description: ID of Application to delete
          type: string
          required: true
        - name: userEmail
          in: query
          description: The email address of the owner to remove from the application
          type: string
          required: true
      responses:
        '200':
          description: Successfully deleted owner
          schema:
            $ref: '#/definitions/applicationType'
        '403':
          description: Not allowed
        '404':
          description: Not found
  '/applications/{appId}/subscriptions':
    get:
      security:
      - accessCode: ['read_subscriptions']
      - implicit: ['read_subscriptions']
      - password: ['read_subscriptions']
      tags:
        - subscriptions
      description: Gets all subscriptions for a given application.
      parameters:
        - name: appId
          in: path
          description: The ID of the application to retrieve all subscriptions for
          type: string
          required: true
      responses:
        '200':
          description: Successful request
          schema:
            title: ArrayOfSubscriptions
            type: array
            items:
              $ref: '#/definitions/subscriptionType'
        '403':
          description: Not allowed
    post:
      security:
      - accessCode: ['write_subscriptions']
      - implicit: ['write_subscriptions']
      - password: ['write_subscriptions']
      tags:
        - subscriptions
      parameters:
        - name: appId
          in: path
          description: The ID of the Application to add an API subscription to
          type: string
          required: true
        - name: subscription
          in: body
          description: The subscription to add
          required: true
          schema:
            $ref: '#/definitions/subscriptionCreateType'
      responses:
        '201':
          description: 'Success, created'
          schema:
            $ref: '#/definitions/subscriptionType'
        '403':
          description: Not allowed
  '/applications/{appId}/subscriptions/{apiId}':
    get:
      security:
      - accessCode: ['read_subscriptions']
      - implicit: ['read_subscriptions']
      - password: ['read_subscriptions']
      tags:
        - subscriptions
      description: Gets a specific API subscription
      parameters:
        - name: appId
          description: The ID of the application for which to retrieve a subscription
          required: true
          in: path
          type: string
        - name: apiId
          description: The ID of the API to retrieve the subscription for
          required: true
          in: path
          type: string
      responses:
        '200':
          description: Successful request
          schema:
            $ref: '#/definitions/subscriptionType'
        '403':
          description: Not allowed
    delete:
      security:
      - accessCode: ['write_subscriptions']
      - implicit: ['write_subscriptions']
      - password: ['write_subscriptions']
      tags:
        - subscriptions
      description: Deletes an application's API subscription
      parameters:
        - name: appId
          description: The ID of the application of which to delete a subscription
          required: true
          in: path
          type: string
        - name: apiId
          description: The ID of the API to delete the subscription for
          required: true
          in: path
          type: string
      responses:
        '204':
          description: 'Successfully deleted, no content'
        '403':
          description: Not allowed
    patch:
      security:
      - accessCode: ['write_subscriptions']
      - implicit: ['write_subscriptions']
      - password: ['write_subscriptions']
      tags:
        - subscriptions
      description: For approving or disapproving subscriptions
      parameters:
        - name: appId
          description: The ID of the application of which to delete a subscription
          required: true
          in: path
          type: string
        - name: apiId
          description: The ID of the API to delete the subscription for
          required: true
          in: path
          type: string
        - name: patchBody
          in: body
          required: true
          schema:
            type: object
            title: SubscriptionPatch
            properties:
              approved:
                type: boolean
              trusted:
                type: boolean
      responses:
        '200':
          description: Succesfully updated
          schema:
            $ref: '#/definitions/subscriptionType'
...
  '/subscriptions/{client_id}':
    get:
      security:
      - accessCode: ['read_subscriptions']
      - implicit: ['read_subscriptions']
      - password: ['read_subscriptions']
      tags:
        - subscriptions
      description: For looking up a subscription by client_id (used for OAuth2 Implicit Grant)
      parameters:
        - name: client_id
          required: true
          description: The client_id created when creating a subscription
          in: path
          type: string
      responses:
        200:
          description: Successful response
          schema:
            type: object
            properties:
              application:
                type: object
              subscription:
                type: object
        404:
          description: Not found.
...